backend/requirements.txt
uvicorn[standard]==0.24.0
anthropic==0.71.0
instructor==0.5.0
requests==2.31.0
networkx==3.2.1
python-multipart==0.0.6
pillow==12.0.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==3.7.1
click==8.3.0
colorama==0.4.6
fastapi==0.120.0
h11==0.16.0
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
pydantic-settings==2.2.1
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0

# Fetch.ai
# uagents==0.4.0
